+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [APR's Version Numbering](https://apr.apache.org/ver
 - [APC-6160]: chrony: Add NTS certificate to chrony.conf
 - [APC-5918](https://jira.iris-sensing.net/browse/APC-5918): Add jsoncpp patch to fix SDK build
 - [APC-6277](https://jira.iris-sensing.net/browse/APC-6277): Allow update via swupdate only if installed firmware has version 2.1.5 or higher
+- [APC-6101]: Add persistent logging support for initramfs's emergency switch
+
 
 ### Changed
 - [APC-5918](https://jira.iris-sensing.net/browse/APC-5918): Update this layer to kirkstone, including refactoring of various recipes